requirements.txt
django>=5.0
djangorestframework>=3.17.1
django-cors-headers>=4.9.0
whitenoise>=6.12.0
pillow>=12.2.0
gunicorn>=25.3.0
psycopg2-binary>=2.9.12
python-decouple>=3.8
